scripts/requirements.txt
markdown>=3.5.1
python-frontmatter>=1.1.0
Pillow>=10.0.0
ascii-magic>=2.3.0
